refactor(quizGameRegistry): split into genericQuizGames, customQuizGames, complexQuizGames (#145)
